refactor: Extract providers into their own file
